requirements.txt
Flask~=2.1.2
psycopg2-binary
